feat(138.a): refactor and move edit to its own component for reuse
